3box comments working in Public Feed
