+++ package.json
@@ -10,12 +10,13 @@
     "postcss-cli": "^7.1.1",
     "react": "^16.11.0",
     "react-dom": "^16.11.0",
+    "react-draggable": "^4.3.1",
     "react-router-dom": "^5.1.2",
     "react-scripts": "^3.4.1",
     "react-spinners": "^0.8.1"
   },
   "scripts": {
-    "build:style":"tailwind build src/styles/index.css -o src/styles/tailwind.css",
+    "build:style": "tailwind build src/styles/index.css -o src/styles/tailwind.css",
     "start": "npm run build:style && react-scripts start",
     "build": "NODE_OPTIONS=--max_old_space_size=4096 react-scripts build",
     "test": "react-scripts test",
