+++ src/styles/landingpage.css
@@ -19,3 +19,31 @@
 
     transition: 0.1s -webkit-filter linear;
 }
+
+.modal {
+  position: absolute;
+  top: 0;
+  left: 0;
+  width:100%;
+  height: 100%;
+  background: rgba(0, 0, 0, 0.6);
+}
+
+.modal-main {
+  position: absolute;
+  background: white;
+  transform: translate(-50%,-50%);
+}
+
+.display-block {
+  display: block;
+}
+
+.display-none {
+  display: none;
+}
+
+
+.threebox-comments-react > .comments_footer {
+  display: none;
+}
